+++ CHANGELOG.md
@@ -5,13 +5,20 @@
 ### Additions
 
 - **[Rock]** Now supports `rock version` - @vknabel
-- **[Rockfile]** Overriding single properties in `Rockfile` are now supported - @vknabel
+- **[Rock]** Now supports `rock run script` - @vknabel
+- **[Rock]** Added convenience scripts for `rock build`, `rock test` and `rock archive` - @vknabel
+- **[Rock]** Derived scripts that will be executed will be highlighted now - @vknabel
+- **[Rockfile]** Every value in `Rockfile` will now be interpreted as Stencil template - @vknabel
+- **[Rockfile]** Overriding single properties in `Rockfile` is now supported - @vknabel
+- **[Other]** Provides experimental `Install.generated.swift` Installation Script - @vknabel
 
 ### Non Breaking Changes
 
+- **[Project]** Now uses krzysztofzablocki/Sourcery - @vknabel
 - **[Project]** Added Swiftlint - @vknabel
 - **[Project]** Added Travis CI - @vknabel
 - **[Project]** Added Danger - @vknabel
+- **[Dependencies]** Uses ColorizeSwift for terminal colors instead of Swiftline - @vknabel
 
 ## 0.2.0